requirements.txt
numpy==1.18.1
pandas==1.0.0
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytz==2019.3
six==1.14.0
tqdm==4.42.0
